Refactor code into Story and DevelopmentTime classes
Improve code readability and maintainability by refactoring development time content script. The DevelopmentTime methods are now in `Story` and `DevelopmentTime` classes. Also, the 'hoursBetweenExcludingWeekends' function was moved to 'utils' directory.
